mod 15964

+++ 20230531/_15964.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main()
 {
-    int n, m;
+    long long n, m;
     cin >> n >> m;
     cout << (n + m) * (n - m);
 }
